BugFix: Ensure CHANGELOG include all commits for release
Per the goreleaser docs, you must set fetch depthness when
checking out the code for GoReleaser to work properly.

The warning in the docs to configure GitHub Actions states:

> Fetch depthness
> Notice the fetch-depth: 0 option on the Checkout workflow
> step. It is required for GoReleaser to work properly.
> Without that, GoReleaser might fail or behave incorrectly.

Ref: https://goreleaser.com/ci/actions/

In addition, there is a common errors page that notes if you
are only seeing one commit, it is likely because GoReleaser was
run against a shallow clone.

Ref https://goreleaser.com/errors/no-history/